requirements.txt
torch
transformers
numpy
webrtcvad
pyaudio
soundfile
sounddevice
gtts
python-dotenv
ollama
